requirements.txt
numpy==1.23.0
tensorflow-cpu>=2.5
torch==1.9.0
einops>=0.3
transformers==4.20.1
wandb
tqdm
unidiff
